Split pager into separate files.
